# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LumoNAS API
  description: |
    LumoNAS management daemon API. Provides endpoints for system monitoring,
    storage management, Docker orchestration, networking, backups, and administration.
  version: 0.1.0
  contact:
    name: LumoNAS
    url: https://github.com/lumonas/lumonas
servers:
  - url: http://127.0.0.1:8080/api/v1
    description: Local development
...
paths:
...
  /users/{id}/2fa/setup:
    post:
      tags: [Users]
      summary: Start two-factor enrollment
      operationId: setupTwoFactor
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Provisioning secret and one-time recovery codes
  /users/{id}/2fa/enable:
    post:
      tags: [Users]
      summary: Confirm two-factor enrollment
      operationId: enableTwoFactor
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, pattern: '^[0-9]{6}$' }
      responses:
        '200':
          description: Two-factor enabled
  /users/{id}/2fa/disable:
    post:
      tags: [Users]
      summary: Disable two-factor authentication
      operationId: disableTwoFactor
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Two-factor disabled
...